data: add helper that can generate/start/stop the snapd service
The snapd.manage script is the interface that the base snaps will
use to start/stop snapd when snapd is run from a base.

Putting the script inside snapd means that we can evolve it without
having to modify any of the bases that use it. If we e.g. use
more systemd units etc.

There will be a coresponding PR for core18 that will use this
script to launch snapd on boot.

Note that this is currently hard to (spread) test because our
current core already has a snapd.{socket,service}. However once
we can bootstrap a core18 system testing this will happen naturally.