image: add workaround for broken store
The store does currently not return "type: base" in the details
endpoint for snaps that have "type: base" set in "meta/snap.yaml".

To workaround this this commit will set "snap_mode=core" for
snaps of "type: application". This is fine for now because
setBootvars() is only called with two snaps: kernel and core/base.