snap: reject more layout locations
This patch adds the following entries to a list of off-limit layouts:
 - /var/lib/snapd/
 - /var/snap/

Those should not be needed for layouts because they interfere with
internal snapd state and state of other snaps, respectively. As such
layouts are designed to aid in porting software and no software has
legitimate reasons to poke into those places.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>