cmd/snap-update-ns: use RCall with SyscallsEqual
This patch does two trivial changes, it switches all unit tests that
used Calls() (which return an slice of strings) to RCalls which returns
a slice of CallResultError structures. At the same time the assertion
used is switched from DeepEquals to SyscallsEqual.

This has the added benefit of checking return values and error values
and, in case something breaks, offers human-readable error messages.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>