snap: don't include newline in hook environment
Currently, both hooks and apps use the global `environment`, if
specified. However, they both use slightly different mechanisms of
interpreting the environment: hooks append a newline, apps do not. The
newline causes problems when the variable is used in the shell.

Fix this issue by creating a new function to evaluate the environment
(that doesn't append a newline), and have both apps and hooks use it.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>