overlord/ifacestate: translate "core" <=> "snapd" as appropriate
This patch adds a mechanism for transparently re-mapping interface
connection peers for both incoming (API) requests and outgoing (state)
requests.

The mechanism is then applied to make it so that every time a connection
request for "snapd" comes in, it is re-mapped to "core" in the state, so
it seems that the connection is to core if one inspects the state alone.
At the same time explicit requests to "core" are re-mapped to to
"snapd", so that it seems that one can connect interfaces to implicit
slots on core, as before.

In both cases this mechanism is only applied if "snapd" is the host of
implicit interfaces.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>