interfaces: generalize writable mimic profile
This patch changes apparmor rules generated for writable mimics.
Currently those rules are precise, that is, they grant the permission to
construct the mimic at a specific directory and nowhere else.

Mimic permissions are handed out as a result of layout and content
interface connections. The permissions are derived from the directory
used by either of the two constructs.

This works fine for some paths but not all, as we learned. As a simple
example, a layout requesting /lib/foo can be easily constructed. The
directory /lib exists in the core snap and a mimic (which is done for
the base directory of the path in question) around /lib is successfully
created at runtime.

This relies on an unexpected assumption that the mimic is always created
in the base directory of the path defined by the content interface or
layout element. As a counter-example consider a layout requesting
/lib/foo/bar. As before /lib exists and the attempt to create /lib/foo
will encounter the EROFS error, triggering a mimic at /lib. The apparmor
permissions however do not have this flexibility since they always allow
only the parent of the path to be transformed into a mimic, here that
would be /lib/foo.

This patch fixes this issue. As a helper we introduce the notion of
assumed prefix depth. The depth refers to the number of directories that
the code can safely assume are present. The prefix varies by
circumstances. For instance when the path must clearly refer to
/snap/some-name/some-revision/ then it an appropriate value is selected
not to hand out permissions to create /snap, /snap/some-name or
/snap/some-name/some-revision.

Using the prefix depth the algorithm effectively constructs a number of
apparmor permissions for increasing path depth, up until the point where
the original path (that designated by either the layout or the content
connection) is reached.

The patch includes changes to the layout code, content interface code
and a trivial tweak to a spread test that now tests this condition
explicitly.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>