interfaces: prefer "snapd" when resolving implicit connections
This patch changes the logic that resolves implicit connect or
disconnect calls. This adds a third case to the existing two
(ubuntu-core and core), with the highest priority.

The idea is that if the repository has the snapd snap then implicit
connection operations will be resolved in favour of that snap over the
other two. This change will be coupled with similar code that prefers to
add the implicit slots to the snapd snap over the core snap, if both are
available.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>