overlord: introduce a gadget-connect task and use it at first boot (#5317)
This introduces a "gadget-connect" task that connects plugs and slots as instructed by the gadget.yaml "connections" stanza. First boot code then uses it.

It is very similar to the "auto-connect" task in many ways.

I chose to encapsulate this in a task because it's much easier to skip/ignore some operations once we can observe the installed snaps, also when we support updating the model assertion and required snaps (or similar) we will need to perform this again.