strutil: support iteration over almost clean paths
This patch changes the path iterator to support otherwise clean (as in
filepath.Clean-wise) that end with a trailing slash. Such paths are
common in apparmor world and supporting them without errors is
convenient.

The path that the iterator uses is always the input path, that is, clean
is only used as a validator. This allows us to keep the assumptions that
the path is not transformed.

The patch also removes a duplicate test for unclean paths.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>