[RFC] asserts: add (optional) kernel-track to model assertion
This allows specifying the kernel track in the model assertion
so that e.g. core18 can pull in a 4.15 kernel when building a
core18 system.

If this gets accepted we need to add code to `snap prepare-image`
to pull the kernel from the right track and also code to snapstate
to ensure the track of the kernel cannot be changed.

This is also discussed in https://forum.snapcraft.io/t/5947