tests: run "arp" tests only if arp is available
On core18 we have no "arp" utility so skip this part of the test
if no "arp" is available. We could use "arp neigh add/del" when
arp is not found but given that the ip tool is already used in
the test to manipulate the network this is probably sufficient.