snap: support hook environment
Currently, snapd supports the `environment` keyword in the root of the
`snap.yaml`, as well as nested under individual `apps`. However, it does
not support them under individual `hooks`. Add support for this, which
lays more of the groundwork necessary for the snapcraft CLI to get rid
of its wrappers.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>