tests: add missing slots in classic and core provider test snaps
Since there may be policy differences when a slot implementation is installed
on classic distro vs all-snaps, it is important that the testsuite generates
policy for both. When doing other work, it was discovered that when creating
the test-snapd-policy-app-provider-classic snap, the fwupd slot was omitted.
Recently, wayland and x11 were updated to be installed as app snaps, but the
test-snapd-policy-app-provider-core snap was not updated.