tests/lib/prepare-restore: update Arch Linux kernel LOCALVERSION handling
Since 2018.07.30 the kernel localversion is set by scripts/setlocalversion in
the kernel tree and no longer matches the package version as seen in `pacman -Qi
linux` output. What before was eg. 4.16.13-2-ARCH with package version 4.16.13-2
is now changed to 4.17.11-arch1, package version 4.17.11-1.

Refactor the check look at the contents of the linux package instead and match
that with the running kernel. This will make us immune to any future package
versioning changes.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>