many: add `snap set core network.rate-limit` option
This adds a new option to rate limit the download of snaps. It is
in RFC state at this time, the idea is that it can be used to
limit the speed of auto-refreshes. It is currently applied to all
downloads though (this will be fixed if this PR has a chance to
survive).