image: detect and error if bases are missing
Combined with the ability to order snaps correctly on firstboot
this will ensure we cannot create seeds that do have missing
bases.

As a next step we should consider to add the bases automatically.
The downside of doing this is that the people building images
may end up with lots of different bases without even knowning
it.