interfaces: add vendor-id: vendor-name map
This patch begins the refactoring of adb-support interface to feature a
built-in USB vendor id to name table. As a part of constructing the
table out of the existing rules I found a duplicate value (0x2207) which
was attributed to Rockchip and Emerson. A quick look at
http://www.linux-usb.org/usb.ids seems to imply that Rockchip is the
right name since Emerson doesn't register any hits there.

This also shows the advantage of this model, that we get automatic
duplicate detector from the compilation process itself.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>