devicestate: error if gadget uses base different from the model.Base
The gadget snap is ordered early in bootstrap before other bases
but after the model base. This means that if a gadget would use
a base different from the model base things may fail so detect
this case and error.

If we find a use-case for a gadget to have a different base
than the model base we can re-evaulate this and order differently.