selftest: detect if apparmor is unusable and error
Under some configuration apparmor may look like its available but
we get permission denied errors when trying to use it. This can
happen on e.g. an lxd container that runs with:

```
raw.lxc: |
  lxc.apparmor.profile=unconfined
```

In this case lxd will not setup apparmor stacking so the container
looks unconfined however lxd will not grant CAP_MAC_ADMIN to the
container (which is quite sensible). But it means that snapd will
not be able to setup any apparmor profiles inside such containers.

When this is detected snapd will refuse to run because we cannot
support this configuration. The host apparmor confinement will
interfere with the container and inside the container we can
not do anything about this. See the unsuccessful PR
https://github.com/snapcore/snapd/pull/5621 for an attempt
for an attempt to support this environment.