daemon,overlord: fix race when going into socket activated mode
Make sure the code deals with changes created while restarting into
socket activation mode. The old code had the following race
condition:
1. snapd starts
2. Ensure runs
3. A `snap install foo` request comes in while Ensure runs
4. In Ensure the system notices that it can go into socket
   activation mode and starts stopping.
5. While stopping the code that handles the pending requests
   create a new change that will install a snap.
6. The system goes into socket activation mode and the pending
   change is not hanlded

The new code adds an extra check in step (6) to see if it really
can go into socket activation mode or if a change was created
in the meantime. If this is the case then snapd will do a "normal"
restart and after the restart the new change will be processed
(the restart is still performed because it is easier than to
undo all the shutdown steps that already happend at this point).

Some extra tests were also added.