daemon: add pokeTest helper to the daemon tests
This ensures that we do not hold the state lock while doing
store operations in the daemon. When adding this it turned
out that a lot of the tests do not setup the store correctly,
they just reply on previous tests having called "replaceStore"
(which is never undone). This PR adds the relevant setup into
the tests too.