systemd: do not run "snapd.snap-repair.service.in on firstboot bootstrap
The snapd.snap-repair.service is run via a timer service. The
unit itself should not be started by snapd when it bootstraps
the system for firstboot. This PR marks the service as such.

This will fix the firstboot problem that we see on the dragonboard
where the snapd snap cannot bootstrap because there is no network
and this makes snap-repair fail.