strutil: fix MatchCounter to also work with buffer reuse
`strutil.MatchCounter` had a bug where, if the caller reused the `[]byte` in
later calls, the contents of a (supposedly) internal buffer got overwritten.

This fixes that by switching to copying it to a bytes.Buffer. We could also
instead do the accounting ourselves but that seems like too much fiddly work.