overlord: add chg.Err() in testUpdateWithAutoconnectRetry
We are currently seeing build failures in the edge PPA with
the following error:
```
FAIL: managers_test.go:2522: mgrsSuite.TestUpdateWithAutoconnectRetrySlotSide

managers_test.go:2523:
    ms.testUpdateWithAutoconnectRetry(c, "some-snap", "other-snap")
managers_test.go:2514:
    c.Assert(chg.Status(), Equals, state.DoneStatus)
... obtained state.Status = 9 ("Error")
... expected state.Status = 4 ("Done")
```
This happens quite frequently but seems to be dependent on the
environment, i.e. it is not reproducible when run locally outside
of the sbuild environment.