cmd/snap: handle "snap interfaces core" better
The legacy "snap interfaces" command does client side filtering and thus
doesn't benefit from the server-side nickname/remapping created when the
core18 snap was introduced. With the introduction of the re-mapping
logic implicit slots are provided by various snaps that are always
exposed, on the API layer, under the "system" nickname.

The re-mapping is done server side so various commands like "connect"
and "disconnect" work perfectly fine with both old and new names. The
"snap interfaes" (plural) command is the exception as it was doing all
of the filtering locally.

This patch adds a tiny special case that allows matching "system"
nickname when the user has requested the "core" snap.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>