overlord/snapshotstate: small refactor of internal helpers
overlord/snapshotstate had an internal helper,
`snapNamesInSnapshotSet`, that used to returned the snap names and
their filenames of a given snap id with matching snaps; it was
`func(uint64, []string) ([]string, []string, error)`, and was used by
restore/check/forget to find the snapshots that were part of a set and
that needed to be included in a given change.

In #5884, that function changed to also return the snap IDs, so that
restore could error out if a snap's ID changed around a restore. it
was, thus, `func(uint64, []string) ([]string, []string, []string, error)`

Next I want to block on Epoch changes... so rather than going with a
`func(uint64, []string) ([]string, []string, []string, []snap.Epoch, error)`
I decided to refactor it a little bit, introducing an internal, ad hoc,
`snapshotSnapSummary`.

I hope you like it.