cmd: expand helper process IPC
This patch changes the architecture of the helper process to allow
sending different kind of IPC requests. The IPC system is updated from
from eventfd to a pipe.  For now we have just two commands: exit and
capture mount namespace, matching existing functionality. Some of the
debugging messages have been shortened and reworded to express the new
nature of the helper. Helper code has moved from a (already too long
function) into a set of smaller functions for readability.

About the helper process:

    The helper process is spawned in the initial mount namespace, before
    snap-confine ventures deep into mount namespace and then per-user
    mount namespace for various initialization steps. It lurks waiting
    to perform the capture which cannot be otherwise done by
    snap-confine's main process. The system ensures that the process
    will die after three seconds without any IPC requests from the
    parent, in case a programming error somehow kills the parent without
    asking the helper to terminate.  This is all exiting functionality,
    this part is just for reference and refreshment.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>