overlord/snapshotstate: some renaming and refactoring to address review concerns
This change renames the maybeRunUser helper to be tarAsUser, and
specializes it to _only_ run tar instead of arbitrary commands. This
makes it that little bit clearer why the fail-open case is Ok, and why
it's not generally useful as-is. Also unifies sudo/runuser variables
to be userWrapper.

Thank you @zyga, I think it is clearer this way.