cmd/snap-exec: use ReadInfoExceptSize
A try-mode snap is represented by a bind mount from whatever the original
location was, to /snap/$SNAP_NAME/$SNAP_REVISION as well as a symlink
from /var/lib/snapd/snap/$SNAP_NAME_$SNAP_REVISION.snap to whatever
the original location was.

In the execution environment the symlink cannot work, in general. The
code loading snap.Info was using the fully generic snap.ReadInfo which,
as one of the features, tried to stat and compute the size of the snap.

Since reading the size of a snap is not very useful for snap-exec and
since it can fail if the symlink is not reachable in the execution
environment, switch to the more limited helper.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>