tests/lib: adjust to changed systemctl behaviour on debian-9
in debian 9, we will *sometimes* but not always get failures in suite
prepare, along the lines of

    + systemctl daemon-reload
    + systemctl restart snapd.socket
    Job for snapd.socket failed.
    See "systemctl status snapd.socket" and "journalctl -xe" for details.

if you go and look, you'll see

    snapd.socket: Socket service snapd.service already active, refusing.
    Failed to listen on Socket activation for snappy daemon.

which means `systemctl restart foo.socket` no longer restarts
`foo.service`.  This sounds like a bug, but while it gets fixed, this
change works around it.