overlord/snapshotstate/backend: survive missing directories
Without this change a missing "common" directory is enough to abort a
snapshot.

I'm not sure how a "common" directory can exist when the revisioned
directory exists (need to chase zyga for that), but it happens. So
this deals with that, and with the reverse situation (where common
exists but no revisined directory does). It also logs a bit more about
what's going on (in DEBUG), to help if things misbehave in the future.