Split nested vm suite on core and classic and new snapshots test
This is split because in the future the setups will be different in case
the nested vm is core or classic.

Also it is included a new test for snapshots which tests in case the
core is refreshed and reverted. Until the snapshots feature is in stable
we need to run from beta to edge.

To run it do: SPREAD_CORE_CHANNEL=beta SPREAD_CORE_REFRESH_CHANNEL=edge
spread
google-nested:ubuntu-16.04-64:tests/nested/classic/snapshots-with-core-refresh-revert

Those tests will be executed on each new core on edge soon as part of a
spread cron job.