cmd/libsnap: add simplified feature flag checker
Prior approach to add a bridge for knowledge from snapd to via the facts
system was declined (though it may be brought back for another reason to
expose snap installation flags to snap-confine). Still, to move forward
I'd like to propose a more basic system where presence of a file in
/var/lib/snapd/experimental is considered as the feature being enabled.

Feature flags files will be written and removed by snapd in the
configuration handler for core. This will be done in a subsequent patch.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>