sanity: refuse to try to do things on WSL.
Without this change, snapd could be started on WSL (daemons won't run,
but snapd can be started by hand). This'll put an end to that.

Also, snap would try to communicate with snapd on WSL, and error out
with the rather unhelpful "no such file or directory" on the socket.

This change puts WSL on parity with darwin: you'll see the help, and
can pack, download, and look at version; anything else should give you
an error message explaining the problem:

```
Interacting with snapd is not yet supported on Windows Subsystem for Linux.
This command has been left available for documentation purposes only.
```