cmd/snap-seccomp: define missing PTRACE_ values
Some ptrace requests are not available on all architectures. Since the
values for those requests just follow a simple sequential enumeration
it should be safe to define the missing values. The kernel doesn't
support them but we don't care about that because we just want to allow
them if they are specified by the seccomp profile.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>