dirs, wrappers, overlord/snapstate: make completion + bases work
The old completion code assumed that if a snap was installed `<snap
mount dir>/core/current/usr/lib/snapd/complete.sh` was there, and used
that to complete things.

This is not true now that you can have a purely core18-based system,
for example, so completers wouldn't get created.

This fixes that: complete.sh is from core if the snap has no base,
from the snapd snap if it is installed, and otherwise from the host.