Skip opensuse from interfaces-openvswitch-support test
Next update for opensuse is failing when interfaces-openvswitch-support
test is executed. The cause is the same than for arch system, where the
interface is allowing access to /run/uuidd/request and in these systems
the request is done in /run/run/uuidd/request, making fail the snaps
which try to request a random id.

test error:
https://paste.ubuntu.com/p/bv9xZj36XR/

debug info:
https://paste.ubuntu.com/p/nMF4BR8ZF7/