cmd/snap-confine: use snap-discard-ns ns to discard stale namespaces.
During startup snap-confine may decide to discard a preserved mount
namespace that became stale. This used to involve unmounting one file.
Then we realized we had a bug because the .fstab file was not unlinked.
With the introduction of per-user mount namespaces we need to do a little
bit more. Instead of duplicating this logic or providing library code
I believe the easiest way is to execute snap-discard-ns itself.

The same patch contains a small refactoring for opening internal tools.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>