client: return error from AppInfosFromSnapAppInfos
This changs `client.AppInfosFromSnapAppInfos` to also return an error,
instead of logging it and carrying on. The error is then returned to
the user (when handling a `services` call) or just logged (when using
it to fill in a snap).

(These errors should be relatively rare and signal systemd being in
trouble, fwiw).