release: detect too old apparmor_parser
This patch re-factors apparmor probing logic so that we can detect older
apparmor parser (with fewer features than required to run snapd in any
way). The API has been streamlined, tests simplified.

The main change is that probing of kernel and parser features happens in
tandem. If some kernel features are present but required parser features
are absent then "level" is set to unusable. This is a new level that
that sits between none and partial.

Most importantly this is the level that is not causing the apparmor
backend to be added to all security backends.

In practical terms this means we run without apparmor on openSUSE Leap
42.3 which in turn fixes a bug that prevented snapd from working
correctly with stricter error reporting during startup phase when system
key is considered.

Extra care has been taken to preserve the property that comparing the system
key does not necessarily need to run apparmor_parser, so that fast path of
"snap run <snap>" remains efficient.

A simple spread test ensures that openSUSE Leap 42.3 is not using
apparmor anymore.

Fixes: https://bugs.launchpad.net/snapd/+bug/1805485
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>