interfaces/apparmor: fail on snap-confine profile errors
When the apparmor backend sets up security for the "core" or "snapd"
snaps it performs additional special setup of the snap-confine security
profile. This operation may fail, for example when apparmor_parser is
killed by systemd or when apparmor_parser cannot handle the syntax we
are using in our profiles.

When this happens we used to just log the error and carry on anyway.
This patch changes this behavior in agreement with the prior change of
the system-key profile re-generation logic.

Note that this patch will surface another error, one where openSUSE
Leap 42.3, using apparmor_parser 2.10.3 cannot compile the profile
for snap-confine at all. That issue is addressed separately and is
tracked at https://bugs.launchpad.net/snapd/+bug/1805485

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>