interfaces/backends: detect too old apparmor_parser (2.36)
This patch changes logic conditionally adding the apparmor backend to not do so
if the kernel has some support in the case that the userspace parser is too old
to parse the profiles we generate.

In practical terms this means we run without apparmor on openSUSE Leap
42.3 which in turn fixes a bug that prevented snapd from working
correctly with stricter error reporting during startup phase when system
key is considered.

A simple spread test ensures that openSUSE Leap 42.3 is not using
apparmor anymore.

Fixes: https://bugs.launchpad.net/snapd/+bug/1805485
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>