systemd: start snapd.autoimport.service in --no-block mode
When seeding in core18 we need to make sure that the
snadpd.autoimport.service is started in non-blocking mode. Otherwise
it would wait forever because it has a "After=snapd.seeding.service"
but the snapd.seeding.service will only finish if the code that
writes the core18 snapd units can run. So this prevents a deadlock.

Note that by starting it with --no-block the right thing will happen,
i.e. the unit is starting but waits until the snapd.seeded.service
is done and then it will run (which is what we want).