wrappers: only restart service in core18 when they are active
This fixes the issue that on the very first boot the
snapd.seeded.service must not be restarted. This service has a
"Requires=snapd.socket" relation and that means that when we
seed and we start snapd.socket for the first time we must not
stop it or otherwise systemd will also stop snapd.seeded.service
which will means that systemd will start the units configured with
"After=snapd.seeded.service" which would be too early.