image: do not write empty etc/cloud
The `snap --prepare-image` command will currently look into the
gadget and create a /etc/cloud/cloud.cfg configuration if the
gadget contains a "cloud.conf" file in cloud-init format. But
even if there is no such file an empty /etc/cloud directory is
created.

This empty directory ends up on /writable/system-data/etc/cloud
and that will block the initial boot from populating that dir
with the data from the core18 /etc/cloud directory when this
dir is in "transition" mode in the /etc/system-image/writable-paths.

As a first step we should not create this /etc/cloud directory if
there is nothing in it. This means we will be able to revert:
https://github.com/snapcore/core18/pull/106

Once that is done we need to also support ds-identify.conf or
switch to a cloud.conf/ directory instead of the current file
only approach. The reason is that we do not want to merge the
cloud config from core18 and from the gadget as this is messy.
Instead it should either come from the gadget or from the core18.