cmd/snap: only auto-enable unicode to a tty
The current behaviour of `--unicode=auto` (the default) is to output
unicode depending on the contents of a few locale-related environment
variables, independenlty of whether stdout was going to a terminal or
not. This was considered a bug.

With this change, the behaviour changes to not output unicode to a
pipe unless `--unicode=always` is given.