many: remove .user-fstab files from /run/snapd/ns
When snapd is being uninstalled or when tests are removing past state we
now need to handle the new .user-fstab files in /run/snapd/ns. Those
represent the actual state of the per-user mount namespace that is
persisted when the corresponding snapd feature is enabled.

The extension matches existing .user-fstab files placed in
/var/lib/snapd/mount, for consistency.

The files were already handled by snap-discard-ns but not by various
helper scripts so here it is.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>