many: make Remodel() download everything first before installing
For remodel we cannot assume that we have network access during
the operation. Some remodeling may switch to a different way of
managing the network or similar. This means we need to order the
wait chains in the remodel task so that all downloads and checks
for assertions happen first and the actual install operations
(that do not require network) happen after this.

This PR implements this by taking the TaskSet from Install() or
Update() and marking the "check-assertions" task with a new
TaskEdge marker. This way devicestate can split the taskset on
this marker and adjust the wait chains accordingly.