tests: when running as autopkgtest do not rebuild the deb
The current prepare/restore code will always rebuild the deb
and then install that for each test.

We do not want this behaviour for the autopkgtest backend because
autopkgtest will use the deb from the archive for the testing
and our rebuild changes the test in subtle ways.

This PR also needs a test to ensure we excercise the actual
codepath and are not running into unexpected autopkgtest failures
at deb upload time.