devicestate: add support to Remodel() to a different kernel
This PR allows to remodel to a different kernel. It will ensure
the kernel matches the new model assertion. It also changes
Remodel() to return a full state.Change() instead of the usual
[]*state.TaskSet(). The reason is that a Remodel() cannot be
composed with other changes and also because we need special
data (like the new model assertion) on the change.