snap-confine: do not init device cgroup if we are in "devmode"
Interfaces that use the device cgroup are currently not usable
in devmode because we do initialize the device cgroup in
snap-confine regardless if the snap is in devmode or not.

This PR fixes it by checking if the snap is in devmode. As there
is no good way today this PR adds an indirect way to detect
devmode by looking at the seccomp profile. In the future we
probably want to revisit this.