bootloader: split lk bootloader into lk/lkImageBuilding
The LK bootloader is first bootloader we have which uses a
disk partition to store its configuration. So far grub/uboot
use a configuration file on a filesystem.

This new design makes it more difficult to map the current
design to the existing bootloader abstractions. One way to
make this nice is to provide two "bootloaders" for LK. One
takes care of the image-building and the other one is for
the runtime part when the bootloader writes directly to a
disk partition.