tests: use "retry-tool" instead of infinite while loop for init
When we wait for system initialization we currently use a infinite
while loop a lot. This is problematic because if anything goes
wrong the test times out. This was brought up in the recent
remodel PRs.

This commit changes the patter to use the retry-tool to  wait
for system initialisation.