snapcraft.yaml: add base and adopt-info, rm builddeb plugin
We can modernize the snapd snap by setting base to core, which allows the snap 
to be built with the modern snapcraft, using multipass or lxd (or natively
with --destructive-mode). This also means we need to drop usage of build in
favor of override-build, since in modern snapcraft the build keyword is removed
entirely.

Also, with a modern snapcraft, we can set the license without using passthrough,
and we also should migrate to using snapcraftctl set-version, which is the 
preferred way to set the version in override-pull.

We can also now eliminate usage of the x-builddeb custom plugin and just use the
same commands as part of the override-build snippet.

The test is now updated to build the snap natively with --destructive-mode, the
lxd provider isn't quite stable enough for us to rely on it here and run the
tests on other OS's.

Unfortunately, we are still not able to set type: snapd yet, since the store
isn't ready, so just add that there as a TODO.

Thanks to @stolowski and @cmatsuoka for some of these changes!

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>