snapstate: do not try up update bootloader in ephemeral mode
When installing/seeding a base/kernel in ephemeral mode there
is no need to update the bootloader because kernel/base will
come from the recovery system and it does not make sense to
try to update the bootloader at this point.

This commit adds code that detects when the system is in
ephemeral mode and will skip setting the bootloader then.