boot: write compat UC16 bootvars in makeBootable20RunMode
This commit writes the old UC16 boot environment variables. With
that we can seed correctly in run mode. This is just a step on
the way to the full UC20 boot where we need to use a static
filename for the "good" kernel and the "try" kernel. But it
unblocks seeding and allows us to move forward with writing
end-to-end tests.