many: use naming.WellKnownSnapID() in tests
This is a followup for PR#8335 that uses the naming.WellKnownSnapID()
in some tests to make it easier to read the tests. It's slightly
wasteful as it adds some fmt.Sprintf() where before it was just
static text but IMO it's worth it because the code feels easier
to read this way (at least to me).