configstate: add support for "virtual" configuration
This commit adds the concept of "virtual" configuration. For things
like the hostname we don't actually want to store the value in
the snapd state. Instead on Get() and Set() we use systemd-hostname
interface to get or set the hostname.

This is done via a new "hijack" functionality for the config,
configstate and configcore which is the implemnted by helpers
inside the configcore.