boot: simplify the TestMakeBootable20RunModeSealKeyErr test
The TestMakeBootable20RunModeSealKeyErr is concerned about testing
that errors from secboot.SealKey() are correctly handled in
boot.MakeBootable(). This means there is no need to test that
all the parameters were passed to the mocked function and that
we have a full setup. This is already tested in the
TestMakeBootable20RunMode test. Hence this commit simplifies the
test to the minimum needed to check that errors from
secboot.SealKey() are correctly passed back.